Update producer.json, to include information on mThreat Technology Inc.
This is the parent organization, for well known brands such as MagicSpam, SpamRats and others.  Especially known for it's IP reputation, analytics, and mitigation tools.  It's RBL's are used worldwide, and known for their accuracy and low false positives.  It actively works with organizations worldwide in identifying new and emerging threat sources.